+++ src/FilamentMediaManagerPlugin.php
@@ -76,6 +76,6 @@ class FilamentMediaManagerPlugin implements Plugin
 
     public static function make(): static
     {
-        return new static;
+        return app(static::class);
     }
 }