+++ src/Http/Controllers/FolderController.php
@@ -2,8 +2,8 @@
 
 namespace TomatoPHP\FilamentMediaManager\Http\Controllers;
 
-use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Controller;
 use TomatoPHP\FilamentMediaManager\Models\Folder;
 
 class FolderController extends Controller